pyproject.toml
[project]
name = "Qshing_server"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "bnbong", email = "bbbong9@gmail.com"},
]
dependencies = [
    "pydantic-settings>=2.7.1",
    "uvicorn[standard]>=0.34.0",
    "fastapi>=0.115.7",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "torchaudio>=2.5.1",
    "transformers>=4.48.1",
    "selenium>=4.28.1",
    "webdriver-manager>=4.0.2",
    "beautifulsoup4>=4.12.3",
    "tqdm>=4.67.1",
    "html2text>=2024.2.26",
    "langdetect>=1.0.9",
    "bs4>=0.0.2",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}

[dependency-groups]
dev = [
    "pre-commit>=4.1.0",
    "mypy>=1.14.1",
    "black>=24.10.0",
    "isort>=5.13.2",
    "pytest>=8.3.4",
]

[tool.pdm]
distribution = false

[tool.mypy]
explicit_package_bases = true
namespace_packages = true
ignore_missing_imports = true
plugins = [
    "pydantic.mypy"
]
disable_error_code = [
    "annotation-unchecked",
]

[[tool.mypy.overrides]]
module = [
    "torch.*",
    "selenium.*",
    "webdriver_manager.*",
    "html2text.*",
    "transformers.*",
    "tqdm.*",
    "langdetect.*"
]
ignore_missing_imports = true
